+++ philo/philo.h
@@ -10,11 +10,11 @@
 
 typedef struct  s_infos
 {
-    unsigned long   nb_philo;
-    unsigned long   time_to_die;
-    unsigned long   time_to_eat;
-    unsigned long   time_to_sleep;
-    unsigned long   nb_times_eat;
+    unsigned int   nb_philo;
+    unsigned int   time_to_die;
+    unsigned int   time_to_eat;
+    unsigned int   time_to_sleep;
+    unsigned int   nb_times_eat;
     struct timeval  tv;
 }               t_infos;
 
@@ -27,14 +27,14 @@ typedef struct  s_philo
 
 typedef struct  s_args
 {
-    unsigned long   *tab;
+    unsigned int   *tab;
     t_philo         *philo;
 }               t_args;
 
 void            args_nbr_error(void);
 void            args_syntax_error(int ac);
 void            allocation_error(void);
-unsigned long   ft_atoi(const char *nptr, int ac);
+unsigned int   ft_atoi(const char *nptr, int ac);
 void            init_t_args(int ac, char **arg, t_args *args);
 void            *start(void *arg);
 void	        init_t_infos(t_infos *ithread, t_args *args, int ac);
